refactor: separate components by project
Seperated components in the UI library by the project they are used by.
