+++ apps/sugaming-site/app/[locale]/page.tsx
@@ -11,7 +11,7 @@ import {
   CardFooter,
   CardHeader,
   CardTitle,
-} from '@sugaming/sugaming-ui';
+} from '@sugaming/sugaming-ui/components/common';
 
 export default function Index() {
   const t = useTranslations('Index');
