+++ libs/sugaming-ui/src/components/common/card/components/card-description.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { cn } from '../../../utils';
+import { cn } from '../../../../utils';
 
 type CardDescriptionProps = React.HTMLAttributes<HTMLDivElement> & {
   className?: string;
